// src/Interfaces/candidateDetails.interface.ts
export interface candidateDetails {

            status: string
            interview_time: string
            interview_date: string
            email: string
            panel: string
            candidate_name: string 
            roll_no: string
            graduation_year: string
            phone_number: string
            feedback: string
            type:string

}